+++ internal/controllers/component_customizer_test.go
