+++ internal/controllers/genericprovider/genericprovider_interfaces.go
