+++ util/util.go
@@ -18,7 +18,7 @@ package util
 
 import (
 	operatorv1 "sigs.k8s.io/cluster-api-operator/api/v1alpha1"
-	"sigs.k8s.io/cluster-api-operator/controllers/genericprovider"
+	"sigs.k8s.io/cluster-api-operator/internal/controllers/genericprovider"
 	clusterctlv1 "sigs.k8s.io/cluster-api/cmd/clusterctl/api/v1alpha3"
 )
